requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.7
face-recognition==1.3.0
opencv-python==4.8.0.74
numpy==2.0.2
